Fix check if user is creator of image
